create menus routes for navigation + footer
